Remove the FormatMode Enum and Bounded instances
These instances were only used in tests; while this is an incompatible
change, it will be bundled with further such changes, in order to
actually simplify the FormatOption/FormatMode situation, and progress
towards fixing #1. Another reason to remove them is that conceptually
these instances make no sense for FormatMode, in contrast to (for
example) Unit.