Incompatible API changes
In order to get rid of the two-level FormatOption/FormatMode
formatting choices, move the "fixed unit" concept to a new constructor
of FormatMode, and get rid of FormatOption entirely.

This breaks API completely with regards to formatting
(formatValue/showValue); I've debated introducing new functions and
deprecating the old ones, but given the usage level (small), I don't
think this is.

This mostly simplifies the tests as well, with one exception: some of
the tests apply only to "auto-formatting", which in a way points that
the old FormatMode (fully automatic) had some rationale.