Further internal cleanup to processUnit
The current result type (Just (Either String Unit)) is
counter-intuitive: both 'Nothing' and 'Just (Right u)' are valid
results, but 'Just (Left e)' is an error. The patch changes the result
type to 'Either String (Maybe Unit)', making all 'Right' results to be
valid, and rewriting the surrounding code a bit to make it more clear,
which it succeeds (IMHO). Code coverage increases a bit.